moved auxiliary modules from "rest" into "helpers"
