+++ backend/aka/rest/rentenota.py
@@ -1,6 +1,6 @@
 from aka.rest.base import JSONRestView
-from aka.rest.prisme import Prisme
-from aka.rest.utils import AKAUtils
+from aka.helpers.prisme import Prisme
+from aka.helpers.utils import AKAUtils
 from django.http import HttpResponse
 from django.http import HttpResponseBadRequest
 from django.conf import settings
